cmd/component-render-manifest: fix panic
Running `lokoctl component render-manifest` w/o a component name
currently leads to a panic. Update it to make the behavior identical to
`lokoctl cluster install`: when a component name is given, render the
manifests only for that component, otherwise render manifests for all
components for which we find lokocfg definitions.

Resolves #147