aws add wildcard DNS RR for ingress
By default, create a wildcard DNS RR `*.<cluster_name>.<dns_zone>` to
have a domain for ingress resources right away after the cluster was
created.

This is needed when we install additional components immediately after
the cluster was bootstrapped and those components need to be reachable
via public ingress endpoints.

Occupying the the whole subdomain maybe is not what every user wants,
i.e. we might approach this differently in the future.

We can do that on AWS, since `<cluster_name>.<dns_zone>` points to a
network loadbalancer that forwards tcp/80 and tcp/443 to the worker
nodes of the cluster.