pkg/config: support loading lokocfg from single file
Currently, lokoctl loads all `.lokocfg` files from the current directory
($PWD) when run. There's no way for users to specify a different
directory or to load only a single file.

The same is true for the `lokocfg.vars` file: it's loaded from
`./lokocfg.vars` with no option to change.

Allow users to specify a path to a directory with `.lokocfg` files or a
single file only. Also allow to point to a different `lokocfg.vars`
file.