install: pack and install only necessary files
Currently we pack entire 'lokomotive-kubernetes' submodule, which
contains a lot of unnesecary files for the end user, like README.md,
.git etc. We should only ship Terraform module instead, to reduce the
amount of files, which end user sees.

This commit changes the paths which we are packing, so we only pack
specific modules. We could even push it further and pack only
'flatcar-linux/kubernetes', but it does not harm to pack the entire
provider.

Closes #194

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>