prometheus-operator: Minimize struct initialization in newComponent
The variables can be used directly in the struct initialisation itself
as they are not reused anywhere else in the function newComponent.
However we allow the variable defaultAlertManagerConfig as an
exception as its value is a multi line string and it helps keep the
struct initialisation clean.

Along with that Go will use zero values if a value is not specified
for an attribute of a struct. As a result setting an attribute to an
empty string for example is redundant.