packet: always restart installer if it fails
If there are some network issues and installer fails to download
the image or GPG signature, we should restart installer and try until we
succeed.

This may increase amount of traffic on flatcar infrastructure if
installers gets stuck and downloads images all the time in a loop, but
that should be acceptable, as soon we will have our images in Packer
natively.

Restarting is limited to sleep 60 seconds between attempts and try no
more than 5 times within an hour.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>