packet: Add a configurable default for reservations ids
This patch just adds support to specify what value will be used for the
hardware_reservation_id if a node is not in the reservation_ids map.

This is particularly useful when you don't really care which reservation
goes to which node, but always need reservations on that worker pool. In
that case, for example, you can do:

```
	reservation_ids_default = "next-available"
```

And it will pick the next reservation available **that matches the
facility and instance type**.

This was previously discussed here:
https://github.com/kinvolk/lokomotive-kubernetes/pull/5#issuecomment-495207900

I said there that I was not sure it was going to be needed and can be
added later if needed. It turns out invidian was right and this is
useful, in fact it is very useful for a client and they have exactly the
use case just described in this commit (need most nodes on a worker pool
to have *some* reservation, they don't care which nodes uses which
resrevation id in particular).

Also, the issue mentioned there by invidian
(https://github.com/cncf/cnf-testbed/issues/215) might be a bug that
stops us from using this. But, luckily, it seems to be already solved.

The bug mentioned by invidian has a comment,
https://github.com/cncf/cnf-testbed/issues/215#issuecomment-484135454,
that points to another issue where it is suggested that the
"next-available" option works as expected.

So, I went and tested this myself. I can see that the tfstate file is
updated with the reservation id (something like "111-sadasd-2sd") for
the instance when using "next-available" in terraform and no changes are
required by terraform on later plans/applies. Furthermore, I couldn't
reproduce none of the issues mentioned there, it really seems they are
fixed in Packet API now. I will update the bug soon.