packet: Force RAID 0 creation with 1 spare disk
Even if it is only one spare disk, it makes sense to create a RAID 0
array because the user might want to use that local disk, it adds no
performance penalty, and it is uniform with other setups where more
disks are involved (from the maintenance point of view).

The other option would be to provide a totally different way to access
the disk (without RAID), but that means more complicated provisioning
code for a use case we are not fans of.

Let's extend the RAID 0 to these cases, that also doesn't expand the
test-matrix, and if there is any problem with the approach we can
re-evaluate later.