packet: Fix RAID device renames on reboot
On reboot, the RAID device is sometimes renamed from
/dev/md/node-local-storage to
/dev/md/node-local-storage/dev/md/node-local-storage_1. This causes
problems as /etc/fstab is not updated and, therefore, not used.

This rename bahaviour is documented on mdadm manpage:

        When  arrays  are  created,  this  host  name will be stored in
        the metadata.  When arrays are assembled using auto-assembly, arrays
        which do not  record  the  correct homehost  name  in  their  metadata
        will  be  assembled using a "foreign" name.  A "foreign" name alway ends
        with  a  digit  string  preceded  by  an  underscore  to differentiate
        it from any possible local name. e.g.  /dev/md/1_1 or /dev/md/home_0.

The hosthome option accepts a special name "any", from mdadm manpage:

        The  special  name  "any" can be used as a wild card.  If an
        array is created with --homehost=any then the name "any" will be stored
        in the array and it can be assembled in the same way on any host.  If an
        array is assembled with this option, then the homehost recorded on the
        array will be ignored.

Although the hosthome by default is the node hostname and it shouldn't
change across reboots, we've seen this renames (seems to change after
the iPXE boot) and using this hosthome setting avoids the rename in all
cases. The root cause to understand why the renames happens is still to
be investigated.

In no case we want to rename the array, as it will cause the fstab
entries to fail and disks will not be moved around to other servers
(this are local disks to the server). So, to reduce impact of any future
bug we add this option that completely avoids the rename in these cases.

Therefore, this patch just adds --homehost option to make sure MD arrays
are never renamed because of homehost changing. And specify the name, to
avoid possible issues in the future, as the name is derived from other
flags too (for example, the homehost flag adds a prefix to the name when
not using the wildcard value "any").