gitmodules, assets, components: Move assets to a separate directory
Moving all the assets into a single directory has a potential of
making it easier to manage them. Maybe not in its current form, as the
only thing that has changed is the paths in the `packr.New`
invocations, but that's a first step.

This is first step to address
https://github.com/kinvolk/lokoctl/issues/243. The commit is based on
the work done in https://github.com/kinvolk/lokoctl/pull/244.