docs: Add defaults to Packet configuration, state required changes
The configuration file for Packet had the assets directory default
to /tmp which means that the asset data is lost after reboot.
It also did not clearly state which variables had to be changed
and what their meaning is.

This adds sane defaults where possible and marks those values
that have to be changed through prefixed comments and additional
hints to change them in the following paragraph.