pkg/install: Make Terraform module source relative
Fixes https://github.com/kinvolk/lokoctl/issues/216.

Since the Terraform module source is always in cluster.tf's parent
directory, we can use a relative path in the `source` field when
calling the module.

`${path.root}` is resolved by Terraform to the absolute path of the
directory which contains the root module. This is probably more
portable than injecting a literal absolute path which is
machine-specific.