*: Replace packr2 with vfsgen
Packr has some problems:

- it encodes the absolute path of the asset in some way in the
  generated file, which makes it hard to be reproducible, because
  every developer stores the project under different path, so for each
  developer the generated assets file differ. For this reason we can't
  put the generated files into the repository.

- boxes are a bit awkward to use, because the path passed to the
  packr.New function are used at a "compile time" by packr to generate
  assets, and at a runtime to know which assets should be used. This
  means we can't make the paths dynamic easily.

- the dynamic paths seems to slow down the startup of the application,
  as packr is doing a lot of IO at startup?

So this commit replaces packr with vfsgen, which seems to have no such
problems. This allows us to put the generated files into the
repository, so we can drop an extra steps during build (no need to get
the generator and no need to generate the files), so we are one step
closer to make it possible to build and install lokoctl with `go get`.

This commits drops mentions of packr from most of the places. The only
place where packr is still mentioned is Makefile, because CI is still
using the `install-packr2` rule. It will be dropped later.