Fix race in cluster verification due to rebooting nodes
Cluster verification failed when a node rebooted during the
status check. The first problem was that retryutil.Retry
would stop retrying when getting the node status failed.
The second problem was that in order to print the node
status, it was fetched again without retrying.

The status check for printing is now embedded into the retry loop which
does the readiness check. The retry loop is now also made robust
against reboots by continuing on error.