Allow to pass AWS credentials as env vars
When specified via files, the credentials will
be stored to disk which has security implications.
It is also not necessary to require the user to point
lokoctl to the AWS credentials file.

Instead of reading the credentials from a file, allow to rely on
the Terraform providers's mechanisms to fetch the credentials
from the environment variables or the default config files.