Discover kubeconfig from asset directory
The correct kubeconfig path had to be manually specified
to install a component or check the cluster health.
This meant to provide the absolute path to the kubeconfig
in the asset directory. Using ~/path/to/kubeconfig did not
work because there was no path expansion.

A new function discovers the kubeconfig automatically by
searching in the cluster configuration for the asset path.
This is the new default behavior if the kubeconfig flag or
environment variable is not set. As last fallback value
the default ~/.kube/config is used. Path expansion is handled
for all cases, so that specifying ~/path/to/kubeconfig works.