components/util: Define types for nodeselector & tolerations
Along with types for nodeselectors and tolerations, this commit also
adds a helper function for rendering tolerations as a JSON string.

This will be helpful in components that support nodeselectors and
tolerations. For example: metallb and rook in the near future. At the
moment there is some code duplication in the metallb package, but that
is not within the scope of this patch set and will be followed up as
future work.