AWS/Packet: Rejoin etcd as fresh member when it fails to join
Recreated controller nodes most times failed to join
the etcd cluster because they got removed or their
state now mismatches (cluster state is "initial").
This is related to an etcd limitation (by design)
described in https://github.com/etcd-io/etcd/issues/5422

Add a rejoin script that is run when etcd fails twice during
a short amount of time. It removes the old reference (because
otherwise it is not allowed to join as fresh member), clears
the data directory (because it may have synced some data if
it still was an active member but then panics because the
state mismatches), and sets the initial cluster state to
"existing". Then it restarts etcd. The script is not run
if the etcd service errors because it is not yet provisioned
or it got killed through "systemctl stop/restart".