Re-add data RAID support for worker nodes
The RAIDs were setup during the installation step
which is not used anymore.

Port the RAID setup code from the deleted worker
installation template. Bugs are fixed and the code is
made more robust against accessing mounted drives.
It also got a bit simplified because lsblk can sort by
itself and the RAID mounting code does not need to scan
for drives but can directly reuse the information from the setup.