components/contour: Add tests for rendering manifests
This commit defines test functions that use a mix of
CamelCase_And_Underscores. The reasoning behind this is that
`TestRenderManifest_WithInstallModeDeployment` is much more easier to
read versus `TestRenderManifestWithInstallModeDeployment`. While the
Golang community uses CameCase strictly, I feel that we should not
feel stiffled by convention when it starts to hamper our own
productivity when dealing with code.

A well known poet has said:

"A foolish consistency is the hobgoblin of little minds".