build: Allow to use container image to generate assets/vendor files
The CI checks if the content of ./vendor matches those
it generates when using the current Go version.
Provide a way to locally build in a container with the
latest Go version. Also provide a way to generate the
assets and vendor files from there to include them in a commit.

The structure of the Docker image changed: the binary is
directly installed to the PATH and built from the vendor files.
The container source folder is exactly the local source folder
and can be bind-mounted at any time into the final image for any
command. The makefile rules to generate assets and vendor files
make use of this. Because the /.cache folder of the container is
owned by root, passing --user to docker was not an option but instead,
the chown approach was taken. The `go mod download` step was removed
because it was too slow and downloading modules is not needed when
vendoring can be used.