packet: Create RAID 0 before kubelet starts
This RAID 0 array should be created and mounted before the kubelet
starts, as pods may need to use the node local storage to run on this
node.

This dependency was lost in this PR:

	https://github.com/kinvolk/lokomotive-kubernetes/pull/28

Before this PR, the RAID 0 array was created when IPXE booting and
therefore already existed when the node was rebooted after
faltcat-install. After this PR there is no IPXE boot so the RAID 0 array
is created in the same boot that the kubelet starts, but we missed to
add that dependency that was implicit before (as different systemd units
were executed on different boots, and therefore the array was always
created when the kubelet was about to start).

This PR is just a cleanup to have the dependency, but note that this
doesn't fix node local storage and the previous commit in this PR is
needed for that.