Makefile: Add target to run end to end tests
Add target `run-e2e-tests` which runs end to end tests. You need to pass
two things to this target the `platform` these tests are targeted to and
the `KUBECONFIG`.

If the user sets a variable `RUN_FROM_CI` to `true` then `KUBECONFIG` is
set to the previously fixed location i.e.
`/root/lokoctl-assets/cluster-assets/auth/kubeconfig`.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>