packet: Fix BGP interpolation dependency
Previously, when adding a node to a pool, the BGP attachments were
marked to be re-created by terraform because they depend on all the
workers id and one was being changed.

Changing the element function for the square bracket notation fixes this
dependency on terraform so each BGP assosiation depends only on the
device it is associated with. I couldn't find any documentation about
this, though, but it works reliably.

In terraform 0.12 it is recommended to always use the square bracket
notation instead of element, except when you need a wrap-around
behaviour:
	https://www.terraform.io/docs/configuration/functions/element.html

This is the only clue I could find about preferring square bracket over
element in the docs.