Remove Install method from Component interface
Currently, each component must implement Install method, which is the
same anyway, as we use helper method for that. The helper method is
using RenderManifest() method of Component interface anyway, so having
Install method to only call the helper seems a little bit redundant.

One argument for keeping it would be to keep the API self-contained, but
I think in this case, we reduce boilerplate code and we separate the
component from being aware of 'kubeconfig', which makes the component
more focused on just the component part.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>