pkg/components/util/helm: add chartFromManifests helper method
Currently we store charts in a various formats, either Go templates, or
plain files or Helm chart stored in assets directory. Common part of all
those manifests is, that we can render them to YAML files, which can be
installed on the cluster. Rendering part is done by the
RenderManifests() method.

In order to install our components as a Helm release, we could either
convert all our components to use Helm chart format, however, that
requires a lot of effort.

The least intrusive alternative, is to use RenderManifests() method and
create dummy charts on the fly, which will be installed on the cluster.

This was, we already use estabilished interface we have for components
output, but we use new method of installing them on the cluster.

While this is far from optimal, as for creating a Helm release, there is
more information needed than just YAML manifests, those limitations can
be mitigated, for example by using hardcoded value for installed chart
version.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>