pkg/components/util/install: add InstallAsRelease method
This commit adds InstallAsRelease method, which takes existing component
format we use and installs it on the given cluster as a Helm release,
into special 'lokomotive-components' namespace.

All resources will be created in the namespace, defined by the component
intself. Only component metadata is stored in 'lokomotive-components'
namespace.

Having all components releases stored in special namespace, allows us to
list them and treat them as a storage for installed components on the
cluster. Users should be informed not to install any other Helm releases
into this namespace.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>