e2e test: Change from contour to envoy Daemonset
The contour comoponent runs as a deployment and does not support two
different modes of running, previously supported viz. Deployment and
Daemonset.

Now contour controller runs as Deployment and envoy runs as Daemonset.
So this commit changes the e2e test which tested contour in daemonset
mode now tests enovy.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>