packet/flatcar-linux: add support for CLC snippets
Both kubernetes and kubernetes/workers module in Packet lack the
ability to specify CLC (Container Linux Configuration) snippets that
are used to modify the instances (or devices in Packet).

We can solve that by following the pattern set by code in other
cloud providers, only adapting it to Packet's structure. The variable
names remain consistent.

Signed-off-by: Martin Polednik <m.polednik@gmail.com>