cli/cmd: implement interactive cluster destroy confirmation
Currently there is not any alternative to the `--confirm` flag to confirm the
destruction of a cluster. This forces the user to always have this flag and it
could happen that the user executes this command by mistake. (For instance
when using Ctrl+R or by running and old command with the up key).

This commit introduces an additional interactive way to ask the user for
confirmation if the `--confirm` flag is not present. This should be the
preferred way by users and the confirmation flag is kept for automation tasks.