Verify that all expected nodes came up
When a worker node did not boot up this was unnoticed
because the Kubernetes itself was asked for the available
nodes and not the lokoctl cluster configuration.

Check that Kubernetes knows the same number of nodes
that we expect to come up as specified in the lokoctl
cluster configuration. Also check the node readiness
in the "health" subcommand.