packet: abstract DNS
Route53 is needed to deploy a cluster in Packet, this is a limitation as some
users don't have an AWS account or use another DNS provider.

This commit removes the specific Route53 logic and abstracts the DNS provider by
using a module that receives a list of DNS entries to be created.

variable "entries" {
  type = list(
    object({
      name    = string
      type    = string
      ttl     = number
      records = list(string)
    })
  )
}

The user should specify the DNS module to be used in their configuration:

module "dns" {
  source = "git::https://github.com/kinvolk/lokomotive-kubernetes//dns/route53/"

  providers = {
    aws = aws.default
  }

  entries = module.controller.dns_entries # entries from controller module
  aws_zone_id = "xxxx"
}