pkg/installer/packet: abstract DNS provider
This commit abstracts the DNS provider and implements support for the route53
and manual DNS modules present in lokomotive-kubernetes.

In the manual case it allows the user to set the DNS entries manually by using
the manual DNS provider present on lokomotive-kubernetes.  lokoctl executes
Terraform apply with the -target option to create the controller and workers
nodes, then asks the user to create the DNS entries and after that checks
if the entries were configured correctly, if yes, Terraform apply is executed
once again to finish the deployment.

The route53 is the default provider in lokomotive-kubernetes, so nothing is done
in that case.