openebs-operator: Install CRD `storagepoolclaims`
OpenEBS component does not ship with any CRD. In its current
implementation the operator installs the CRDs. And since operator
installs those CRDs it takes time for those CRD to be available. The
time needed for those CRDs to be available is contingent on:

- Operator pod to be scheduled
- Operator pod image to be pulled
- Operator pod start
- CRDs to be installed
- CRDs to be reconciled

And due to this lokoctl CI fails to install the `storagepoolclaims` CR
and hence prometheus-operator cannot be tested.

Once lokoctl supports helm3 installation and once upstream moves to
helm3 version of chart we can consume that and remove current way of
installation.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>