update terraform render bootkube: Fix the search string
Previously this was using search string for `sed` as "source =", but
this commit changes that to be "source" only. There is a possibility
that `terraform fmt` might format the terraform code and put the equal
sign "=" more than one space away from `source` string to format it
according to the strings after it.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>