pkg/components/util: refactor chartFromManifests
This commit adds replacement function chartFromComponent, which glues
the logic between chartFromManifests and Component interface, making
chartFromManifests to only handle generating the chart.

chartFromComponent also validates, that generated chart is correct,
which makes it safer for the caller.

This commit also adds some unit tests testing chartFromManifests, as now
it is easier to test it.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>