packet: fix DNS entries output
Create a null_resource to force the name of the domains to be computed,
otherwise they will be null on the output.

Also fix the documentation for the user to get the list of entries from the
null resource as the output from internal modules is nost saved on the state
on Terraform 0.12.