Refactor Executor to make it more useful
Refactors how the Executor works in order to remove a few wrapper functions.

The `cluster install` and `cluster destroy` commands now
create an Executor that is passed to the platform. The
platform then uses this executor rather than wrapper functions in order to run their commands.

The `execute()` function is now exported as `ExecuteAsync()`.
A new wrapper, `Execute()`, has been added as
a high level exported function that takes the arguments
and returns an `error` if one occurs, printing the output
from terraform.

In addition, the terraform pkg now has a `Config` struct that is used when calling `NewExecutor`. This struct houses
the WorkingDir/executionPath, and is where future command
line arguments or other configuration will go.