pkg/terraform: check for status in Execute
This commit improves error checking for Execute() function.

Previously, function was mistakenly trying to tail Terraform logs
regardless of whether the Terraform command get scheduled succesfully or
not.

Additionally, it was returning scheduling error instead of the execution
result.

This commit removes error checking from tailing code, by moving it to
separate function. tailFile() function also takes care of stopping the
tailing, when process has been stopped, to avoid leaving orphran
goroutine tailing the file.

If setting up tailing fails, we stil wait for
execution to finish and only then return to ther user. This is to avoid
leaving orphran process running in the background.

Additionally, after the execution is finished, we obtain the result of
it via Status() method and then evaluate it's results, returning
specific error to the user, if needed.

Closes #486

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>