packet/flatcar-linux/kubernetes: add etcd records at the end
Otherwise, when adding controller node, all records are shifted by one,
which triggers re-creating them, which may cause the downtime and
currently produces conflicting records, which breaks adding controller
nodes.

With this commit, etcd records will be added at the end of the
collection and first two entries should always remain the same and be
updated in place.

Closes #181

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>