pkg/terraform/executor.go: don't initialize on actions
As printed output from 'terraform init' it a bit annoying if printed
several times to the user, let's shift the responsibility of calling
Init() to the caller to minimize the number of times we call it.

In the future, we may think about silencing it's output further and
maybe adding some safety checks if Terraform has been initialized (e.g.
by having unexported "initialized" type with exported interface).

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>