aws: copy kubelet kubeconfig to all controllers
Instead of just first one, which is used for bootstrapping.

Right now kubelet on other controller nodes does not start at all, as
the kubeconfig file is missing, which means that HA controlplane on AWS
is broken since the regression introduced in
48f7a727c5c9798fab26bceb99774e61d7065b30.

This commit moves kubeconfig provisioning from bootkube-start to
copy-controller-secrets step, so it will be done for every controller
node.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>