aws workers: don't recreate the pool when launch configuration changes
So changing hyperkube image does not destroy the worker pool, which may
result in loss of drives.

Originally in Typhoon, nodes were treated as immutable objects, so any
changes to it (e.g. changing hyperkube image tag for running kubelet)
would recreate all nodes.

In Lokomotive however, we run kubelet as Kubernetes pod, so upgrades are
done by upgrading Helm release which manages the kubelet DameonSet. That
means we should not require destroying nodes to perform an upgrade.

Closes #48

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>